+++ beacon-chain/sync/subscriber_beacon_blocks.go
@@ -12,6 +12,7 @@ import (
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/blocks"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/interfaces"
 	"github.com/prysmaticlabs/prysm/v5/io/file"
+	"github.com/prysmaticlabs/prysm/v5/time/slots"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -33,6 +34,8 @@ func (s *Service) beaconBlockSubscriber(ctx context.Context, msg proto.Message) 
 		return err
 	}
 
+	go s.reconstructAndBroadcastBlobs(ctx, signed)
+
 	if err := s.cfg.chain.ReceiveBlock(ctx, signed, root, nil); err != nil {
 		if blockchain.IsInvalidBlock(err) {
 			r := blockchain.InvalidBlockRoot(err)
@@ -53,6 +56,79 @@ func (s *Service) beaconBlockSubscriber(ctx context.Context, msg proto.Message) 
 		return err
 	}
 	return err
+}
+
+// reconstructAndBroadcastBlobs processes and broadcasts blob sidecars for a given beacon block.
+// This function reconstructs the blob sidecars from the EL using the block's KZG commitments,
+// broadcasts the reconstructed blobs over P2P, and saves them into the blob storage.
+func (s *Service) reconstructAndBroadcastBlobs(ctx context.Context, block interfaces.ReadOnlySignedBeaconBlock) {
+	startTime, err := slots.ToTime(uint64(s.cfg.chain.GenesisTime().Unix()), block.Block().Slot())
+	if err != nil {
+		log.WithError(err).Error("Failed to convert slot to time")
+	}
+
+	blockRoot, err := block.Block().HashTreeRoot()
+	if err != nil {
+		log.WithError(err).Error("Failed to calculate block root")
+		return
+	}
+
+	if s.cfg.blobStorage == nil {
+		return
+	}
+	indices, err := s.cfg.blobStorage.Indices(blockRoot)
+	if err != nil {
+		log.WithError(err).Error("Failed to retrieve indices for block")
+		return
+	}
+	for _, index := range indices {
+		if index {
+			blobExistedInDBTotal.Inc()
+		}
+	}
+
+	// Reconstruct blob sidecars from the EL
+	blobSidecars, err := s.cfg.executionReconstructor.ReconstructBlobSidecars(ctx, block, blockRoot, indices[:])
+	if err != nil {
+		log.WithError(err).Error("Failed to reconstruct blob sidecars")
+		return
+	}
+	if len(blobSidecars) == 0 {
+		return
+	}
+
+	// Refresh indices as new blobs may have been added to the db
+	indices, err = s.cfg.blobStorage.Indices(blockRoot)
+	if err != nil {
+		log.WithError(err).Error("Failed to retrieve indices for block")
+		return
+	}
+
+	// Broadcast blob sidecars first than save them to the db
+	for _, sidecar := range blobSidecars {
+		if sidecar.Index >= uint64(len(indices)) || indices[sidecar.Index] {
+			continue
+		}
+		if err := s.cfg.p2p.BroadcastBlob(ctx, sidecar.Index, sidecar.BlobSidecar); err != nil {
+			log.WithFields(blobFields(sidecar.ROBlob)).WithError(err).Error("Failed to broadcast blob sidecar")
+		}
+	}
+
+	for _, sidecar := range blobSidecars {
+		if sidecar.Index >= uint64(len(indices)) || indices[sidecar.Index] {
+			blobExistedInDBTotal.Inc()
+			continue
+		}
+		if err := s.subscribeBlob(ctx, sidecar); err != nil {
+			log.WithFields(blobFields(sidecar.ROBlob)).WithError(err).Error("Failed to receive blob")
+			continue
+		}
+
+		blobRecoveredFromELTotal.Inc()
+		fields := blobFields(sidecar.ROBlob)
+		fields["sinceSlotStartTime"] = s.cfg.clock.Now().Sub(startTime)
+		log.WithFields(fields).Debug("Processed blob sidecar from EL")
+	}
 }
 
 // WriteInvalidBlockToDisk as a block ssz. Writes to temp directory.